Refactor code organization
The index.js file is now the main file. A index.d.ts was created.
The index.js is now in the root folder, and the code was moved from the
src folder to lib.
The lib and dist folders are exported, as well as the README.md file,
index.js and index.d.ts.